move classes into ui.
