+++ src/main/java/seedu/address/Main.java
@@ -1,4 +1,4 @@
-package seedu.address.person;
+package seedu.address;
 
 import javafx.application.Application;
