+++ src/main/java/seedu/address/person/commons/util/AppUtil.java
@@ -3,7 +3,7 @@ package seedu.address.person.commons.util;
 import static java.util.Objects.requireNonNull;
 
 import javafx.scene.image.Image;
-import seedu.address.person.MainApp;
+import seedu.address.MainApp;
 
 /**
  * A container for App specific utility functions
