+++ src/main/java/seedu/address/ui/CommandBox.java
@@ -1,4 +1,4 @@
-package seedu.address.person.ui;
+package seedu.address.ui;
 
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
