+++ src/main/java/seedu/address/ui/HelpWindow.java
@@ -1,4 +1,4 @@
-package seedu.address.person.ui;
+package seedu.address.ui;
 
 import java.util.logging.Logger;
