+++ src/main/java/seedu/address/ui/Home.java
@@ -9,4 +9,6 @@ public class Home extends UiPart<Region> {
     public Home () {
         super(FXML);
     }
+
+
 }
