+++ src/main/java/seedu/address/ui/ResultDisplay.java
@@ -1,10 +1,10 @@
-package seedu.address.person.ui;
-
-import static java.util.Objects.requireNonNull;
+package seedu.address.ui;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.TextArea;
 import javafx.scene.layout.Region;
+
+import static java.util.Objects.requireNonNull;
 
 /**
  * A ui for the status bar that is displayed at the header of the application.
