+++ src/main/java/seedu/address/ui/Ui.java
@@ -1,4 +1,4 @@
-package seedu.address.person.ui;
+package seedu.address.ui;
 
 import javafx.stage.Stage;
