+++ src/main/java/seedu/address/ui/UiManager.java
@@ -1,4 +1,4 @@
-package seedu.address.person.ui;
+package seedu.address.ui;
 
 import java.util.logging.Logger;
 
@@ -7,7 +7,7 @@ import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
-import seedu.address.person.MainApp;
+import seedu.address.MainApp;
 import seedu.address.person.commons.core.LogsCenter;
 import seedu.address.person.commons.util.StringUtil;
 import seedu.address.person.logic.Logic;
