+++ src/main/java/seedu/address/ui/UiPart.java
@@ -1,12 +1,11 @@
-package seedu.address.person.ui;
-
-import static java.util.Objects.requireNonNull;
+package seedu.address.ui;
 
 import java.io.IOException;
 import java.net.URL;
-
 import javafx.fxml.FXMLLoader;
-import seedu.address.person.MainApp;
+import seedu.address.MainApp;
+
+import static java.util.Objects.requireNonNull;
 
 /**
  * Represents a distinct part of the UI. e.g. Windows, dialogs, panels, status bars, etc.
